Refactor aml_image_packer.py: - move main class definition to separate file inside lib
