+++ ubt.py
@@ -12,7 +12,7 @@ import sys
 from enum import Enum
 
 from adnl import do_adnl_burn
-from aml_image_packer import AmlImagePack
+from pyamlboot.amlimage import AmlImagePack
 from optimus import do_optimus_burn
